requirements.txt
selenium
twitter
pyyaml